arm64: DT: SM8150: Solve all possible DTC warnings
Solve all the possible DTC warnings across SM8150 device-trees.

The only one that is left is the alias_paths warning and we are
going to silence it because, unfortunately, I have no idea if,
when and why any userspace component (will) want to use these
devices by using their alias.
Pity.

Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>